lite/requirements.txt
fastapi
uvicorn
gradio
python-dotenv
chromadb
requests
APScheduler
pydantic>=2
pyarrow
pandas
pytest
